refactor(racingcar): rename RacingGame -> RacingGameController
