+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,9 @@
 package racingcar;
 
+import racingcar.controller.RacingGameController;
+
 public class Application {
     public static void main(String[] args) {
-        RacingGame.play();
+        RacingGameController.play();
     }
 }
